constructing a global model for a hierarchical graph
